+++ src/Flippa.js
@@ -12,6 +12,7 @@ import Sessions from "./resources/Sessions";
 import SupportEnquiries from "./resources/SupportEnquiries";
 import User from "./resources/User";
 import Users from "./resources/Users";
+import PartnerPage from "./resources/PartnerPage";
 
 import Promise from "bluebird";
 
@@ -89,5 +90,9 @@ export default class Flippa {
 
   get lead() {
     return new Lead(this.client);
+  }
+
+  partnerPage(pageName) {
+    return new PartnerPage(this.client, pageName);
   }
 };