+++ test/unit/Flippa_test.js
@@ -14,6 +14,7 @@ import Sessions from "../../src/resources/Sessions";
 import SupportEnquiries from "../../src/resources/SupportEnquiries";
 import User from "../../src/resources/User";
 import Users from "../../src/resources/Users";
+import PartnerPage from "../../src/resources/PartnerPage";
 
 const expect = chai.expect;
 
@@ -130,6 +131,15 @@ describe("Flippa", () => {
       const flippa = new Flippa();
 
       expect(flippa.lead).to.be.an.instanceOf(Lead);
+    });
+  });
+
+  describe("PartnerPage", () => {
+    it("returns a new PartnerPage", () => {
+      const flippa = new Flippa();
+
+      expect(flippa.partnerPage("domainHoldings")).to.be.an.instanceOf(PartnerPage);
+      expect(flippa.partnerPage("domainHoldings").pageName).to.eq("domainHoldings");
     });
   });
 });